refactor: TranscriptTableコンポーネントをapp/listへ移動
